+++ driver-core/src/test/java/com/datastax/driver/core/Assertions.java
@@ -7,4 +7,8 @@ public class Assertions extends org.assertj.core.api.Assertions{
     public static ClusterAssert assertThat(Cluster cluster) {
         return new ClusterAssert(cluster);
     }
+
+    public static SessionAssert assertThat(Session session) {
+        return new SessionAssert(session);
+    }
 }